Log number of leaked goroutines
